backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
python-multipart==0.0.12

SQLAlchemy==2.0.36
aiosqlite==0.20.0

celery==5.4.0
kombu==5.4.2

httpx==0.27.2

requests==2.32.3

faster-whisper==1.1.0

